added the necessary textfield into the conservation form
