+++ public/static/locales/en/manageProjects.json
@@ -4,6 +4,7 @@
   "detailedAnalysis": "Detailed Analysis",
   "projectSites": "Project Sites",
   "projectSpending": "Project Spending",
+  "protectionStartedIN": "Protection started in",
   "review": "Review",
   "name": "Project Name",
   "nameValidation": "Please enter Project Name",
@@ -14,7 +15,7 @@
   "countTarget": "Tree Target",
   "countTargetValidation": "Please enter Tree target",
   "countTargetValidation2": "Tree target should be more than 1",
-  "conservationImpacts":"Conservation Impacts",
+  "conservationImpacts": "Conservation Impacts",
   "website": "Website",
   "websiteValidationRequired": "Please enter website URL",
   "websiteValidationInvalid": "Invalid website URL",
@@ -166,6 +167,7 @@
   "restorationProject": "Restoration Project",
   "addProjetDescription": "You're few steps away from showcasing your work on Plant-for-the-planet platform.",
   "addProjetContact": "If you need any help contact us at ",
+  "areaProtectedInfo": "Add Some Information",
   "supportLink": "support@plant-for-the-planet.org",
   "wrongCoordinates": "Wrong Coordinates",
   "actionInfo": "The actions your organization is taking to conserve the ecosystem.",
