+++ src/features/user/ManageProjects/StepForm.module.scss
@@ -12,7 +12,7 @@
 }
 
 .stepContainer {
-  margin-top: -50px;
+  margin-top: -60px;
   margin-left: 100px;
   font-family: $primaryFontFamily;
   display: flex;
