+++ src/features/user/ManageProjects/components/DetailedAnalysis.tsx
@@ -1,14 +1,14 @@
 import React, { ReactElement } from 'react';
 import MaterialTextField from '../../../common/InputTypes/MaterialTextField';
 import { useForm, Controller } from 'react-hook-form';
-import i18next from './../../../../../i18n';
+import i18next from '../../../../../i18n';
 import styles from './../StepForm.module.scss';
 import BackArrow from '../../../../../public/assets/images/icons/headerIcons/BackArrow';
 import DateFnsUtils from '@date-io/date-fns';
 import { DatePicker, MuiPickersUtilsProvider } from '@material-ui/pickers';
 import { ThemeProvider } from '@material-ui/styles';
 import ProjectCertificates from './ProjectCertificates';
-import InfoIcon from './../../../../../public/assets/images/icons/manageProjects/Info';
+import InfoIcon from '../../../../../public/assets/images/icons/manageProjects/Info';
 import { putAuthenticatedRequest } from '../../../../utils/apiRequests/api';
 import { localeMapForDate } from '../../../../utils/language/getLanguageName';
 import materialTheme from '../../../../theme/themeStyles';
@@ -100,14 +100,15 @@ export default function DetailedAnalysis({
     { id: 10, title: ready ? t('common:november') : '', isSet: false },
     { id: 11, title: ready ? t('common:december') : '', isSet: false },
   ]);
-  const ownershipType = [
+
+  const enSpecies = [
     {
-      label: ready ? t('manageProjects:tenure') : '',
-      value: 'tenure',
+      id: 1,
+      label: 'animal 1',
     },
     {
-      label: ready ? t('manageProjects:rent') : '',
-      value: 'rent',
+      id: 2,
+      label: 'animal 2',
     },
   ];
 
@@ -191,7 +192,6 @@ export default function DetailedAnalysis({
             plantingDensity: data.plantingDensity,
             employeesCount: data.employeesCount,
             mainChallenge: data.mainChallenge,
-            motivation: data.motivation,
             siteOwnerType: owners,
             siteOwnerName: data.siteOwnerName,
             acquisitionYear: data.acquisitionYear.getFullYear(),
@@ -209,7 +209,6 @@ export default function DetailedAnalysis({
               actions: data.actions,
               activitySeasons: months,
               mainChallenge: data.mainChallenge,
-              motivation: data.motivation,
               longTermPlan: data.longTermPlan,
               landOwnershipType: owners,
               ownershipType: data.ownershipType,
@@ -324,13 +323,17 @@ export default function DetailedAnalysis({
     }
   }, [projectDetails]);
   return ready ? (
-    <div className={styles.stepContainer}>
+    <div className={styles.stepContainer} style={{ marginLeft: '10px' }}>
+      {' '}
       <form
-        onSubmit={(e) => {
+        onClick={(e) => {
           e.preventDefault();
         }}
       >
-        <div className={`${isUploadingData ? styles.shallowOpacity : ''}`}>
+        <div
+          className={`${isUploadingData ? styles.shallowOpacity : ''}`}
+          style={{ marginLeft: '80px' }}
+        >
           {purpose === 'trees' ? (
             <div className={styles.formField}>
               <div
@@ -387,6 +390,7 @@ export default function DetailedAnalysis({
                   </div>
                 </div>
               </div>
+              <div style={{ width: '20px' }}></div>
               <div className={styles.formFieldHalf}>
                 <ThemeProvider theme={materialTheme}>
                   <MuiPickersUtilsProvider
@@ -422,23 +426,6 @@ export default function DetailedAnalysis({
             </div>
           ) : (
             <div className={styles.formField}>
-              <div
-                className={styles.formFieldHalf}
-                style={{ position: 'relative' }}
-              >
-                <MaterialTextField
-                  label={t('manageProjects:projectLocation')}
-                  variant="outlined"
-                  name="projectLocation"
-                  multiline
-                  inputRef={register({
-                    maxLength: {
-                      value: 300,
-                      message: t('manageProjects:max300Chars'),
-                    },
-                  })}
-                />
-              </div>
               <div className={styles.formFieldHalf}>
                 <MaterialTextField
                   inputRef={register({
@@ -454,30 +441,219 @@ export default function DetailedAnalysis({
                   onInput={(e) => {
                     e.target.value = e.target.value.replace(/[^0-9]./g, '');
                   }}
-                  InputProps={{
-                    endAdornment: (
-                      <p
-                        className={styles.inputEndAdornment}
-                        style={{
-                          marginLeft: '4px',
-                          width: '100%',
-                          textAlign: 'right',
-                          fontSize: '14px',
-                        }}
-                      >
-                        {t('manageProjects:hector')}
-                      </p>
-                    ),
-                  }}
                 />
                 {errors.areaProtected && (
                   <span className={styles.formErrors}>
                     {errors.areaProtected.message}
                   </span>
                 )}
+                <div
+                  style={{
+                    position: 'absolute',
+                    top: '-9px',
+                    right: '16px',
+                    width: 'fit-content',
+                  }}
+                >
+                  <div className={styles.popover}>
+                    <InfoIcon />
+                    <div
+                      className={styles.popoverContent}
+                      style={{ left: '-290px' }}
+                    >
+                      <p>{t('manageProjects:areaProtectedInfo')}</p>
+                    </div>
+                  </div>
+                </div>
+              </div>
+              <div style={{ width: '20px' }}></div>
+              <div className={styles.formFieldHalf}>
+                <ThemeProvider theme={materialTheme}>
+                  <MuiPickersUtilsProvider
+                    utils={DateFnsUtils}
+                    locale={
+                      localeMapForDate[userLang]
+                        ? localeMapForDate[userLang]
+                        : localeMapForDate['en']
+                    }
+                  >
+                    <Controller
+                      render={(properties) => (
+                        <DatePicker
+                          inputRef={register({
+                            required: {
+                              value: true,
+                              message: t(
+                                'manageProjects:employeeCountValidation'
+                              ),
+                            },
+                          })}
+                          label={t('manageProjects:protectionStartedIN')}
+                          value={properties.value}
+                          onChange={properties.onChange}
+                          inputVariant="outlined"
+                          TextFieldComponent={MaterialTextField}
+                          autoOk
+                          disableFuture
+                          minDate={new Date(new Date().setFullYear(1950))}
+                          views={['year']}
+                          maxDate={new Date()}
+                        />
+                      )}
+                      name={'acquisitionYear'}
+                      control={control}
+                      rules={{
+                        required: t('manageProjects:ownershipTypeValidation'),
+                      }}
+                      // defaultValue=""
+                    />
+                    {errors.startingProtectionYear && (
+                      <span className={styles.formErrors}>
+                        {errors.startingProtectionYear.message}
+                      </span>
+                    )}
+                  </MuiPickersUtilsProvider>
+                </ThemeProvider>
               </div>
             </div>
           )}
+
+          <div className={styles.formField}>
+            <div
+              className={styles.formFieldHalf}
+              style={{ position: 'relative' }}
+            >
+              <MaterialTextField
+                inputRef={register({
+                  required: {
+                    value: true,
+                    message: t('manageProjects:employeeCountValidation'),
+                  },
+                  validate: (value) => parseInt(value, 10) > 0,
+                })}
+                label={t('manageProjects:employeeCount')}
+                variant="outlined"
+                name="employeeCount"
+                onInput={(e) => {
+                  e.target.value = e.target.value.replace(/[^0-9]./g, '');
+                }}
+              />
+              {errors.employeeCount && (
+                <span className={styles.formErrors}>
+                  {errors.employeeCount.message}
+                </span>
+              )}
+              <div
+                style={{
+                  position: 'absolute',
+                  top: '-9px',
+                  right: '16px',
+                  width: 'fit-content',
+                }}
+              >
+                <div className={styles.popover}>
+                  <InfoIcon />
+                  <div
+                    className={styles.popoverContent}
+                    style={{ left: '-290px' }}
+                  >
+                    <p>{t('manageProjects:employeesCountInfo')}</p>
+                  </div>
+                </div>
+              </div>
+            </div>
+            <div style={{ width: '20px' }}></div>
+            <div className={styles.formFieldHalf}>
+              <ThemeProvider theme={materialTheme}>
+                <MuiPickersUtilsProvider
+                  utils={DateFnsUtils}
+                  locale={
+                    localeMapForDate[userLang]
+                      ? localeMapForDate[userLang]
+                      : localeMapForDate['en']
+                  }
+                >
+                  <Controller
+                    render={(properties) => (
+                      <DatePicker
+                        inputRef={register({
+                          required: {
+                            value: true,
+                            message: t(
+                              'manageProjects:employeeCountValidation'
+                            ),
+                          },
+                        })}
+                        label={t('manageProjects:acquisitionYear')}
+                        value={properties.value}
+                        onChange={properties.onChange}
+                        inputVariant="outlined"
+                        TextFieldComponent={MaterialTextField}
+                        autoOk
+                        disableFuture
+                        minDate={new Date(new Date().setFullYear(1950))}
+                        views={['year']}
+                        maxDate={new Date()}
+                      />
+                    )}
+                    name={'acquisitionYear'}
+                    control={control}
+                    rules={{
+                      required: t('manageProjects:ownershipTypeValidation'),
+                    }}
+                    // defaultValue=""
+                  />
+                  {errors.startingProtectionYear && (
+                    <span className={styles.formErrors}>
+                      {errors.startingProtectionYear.message}
+                    </span>
+                  )}
+                </MuiPickersUtilsProvider>
+              </ThemeProvider>
+            </div>
+          </div>
+
+          <div className={styles.formFieldLarge}>
+            <div className={styles.plantingSeasons}>
+              <p className={styles.plantingSeasonsLabel}>
+                {' '}
+                {purpose === 'trees'
+                  ? t('manageProjects:plantingSeasons')
+                  : t('manageProjects:timePeriod')}{' '}
+              </p>
+              {plantingSeasons.map((month) => {
+                return (
+                  <div
+                    className={styles.multiSelectInput}
+                    key={month.id}
+                    onClick={() => handleSetPlantingSeasons(month.id)}
+                  >
+                    <div
+                      className={`${styles.multiSelectInputCheck} ${
+                        month.isSet ? styles.multiSelectInputCheckTrue : ''
+                      }`}
+                    >
+                      <svg
+                        xmlns="http://www.w3.org/2000/svg"
+                        width="13.02"
+                        height="9.709"
+                        viewBox="0 0 13.02 9.709"
+                      >
+                        <path
+                          id="check-solid"
+                          d="M4.422,74.617.191,70.385a.651.651,0,0,1,0-.921l.921-.921a.651.651,0,0,1,.921,0l2.851,2.85,6.105-6.105a.651.651,0,0,1,.921,0l.921.921a.651.651,0,0,1,0,.921L5.343,74.617a.651.651,0,0,1-.921,0Z"
+                          transform="translate(0 -65.098)"
+                          fill="#fff"
+                        />
+                      </svg>
+                    </div>
+                    <p style={{ color: 'var(--dark)' }}>{month.title}</p>
+                  </div>
+                );
+              })}
+            </div>
+          </div>
+
           {purpose === 'trees' ? (
             <div className={styles.formField}>
               <div
@@ -522,192 +698,6 @@ export default function DetailedAnalysis({
                 )}
               </div>
               <div style={{ width: '20px' }}></div>
-              <div
-                className={styles.formFieldHalf}
-                style={{ position: 'relative' }}
-                data-test-id="employeeCount"
-              >
-                <MaterialTextField
-                  inputRef={register({
-                    required: {
-                      value: true,
-                      message: t('manageProjects:employeeCountValidation'),
-                    },
-                    validate: (value) => parseInt(value, 10) > 0,
-                  })}
-                  label={t('manageProjects:employeeCount')}
-                  variant="outlined"
-                  name="employeesCount"
-                  onInput={(e) => {
-                    e.target.value = e.target.value.replace(/[^0-9]./g, '');
-                  }}
-                />
-                {errors.employeesCount && (
-                  <span className={styles.formErrors}>
-                    {errors.employeesCount.message}
-                  </span>
-                )}
-                <div
-                  style={{
-                    position: 'absolute',
-                    top: '-9px',
-                    right: '16px',
-                    width: 'fit-content',
-                  }}
-                >
-                  <div className={styles.popover}>
-                    <InfoIcon />
-                    <div
-                      className={styles.popoverContent}
-                      style={{ left: '-290px' }}
-                    >
-                      <p>{t('manageProjects:employeesCountInfo')}</p>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-          ) : (
-            <div
-              className={styles.formFieldLarge}
-              style={{ position: 'relative' }}
-            >
-              <MaterialTextField
-                inputRef={register({
-                  maxLength: {
-                    value: 300,
-                    message: t('manageProjects:max300Chars'),
-                  },
-                })}
-                label={t('manageProjects:motivation')}
-                variant="outlined"
-                name="motivation"
-                multiline
-              />
-
-              <div
-                style={{
-                  position: 'absolute',
-                  top: '-9px',
-                  right: '16px',
-                  width: 'fit-content',
-                }}
-              >
-                <div className={styles.popover}>
-                  <InfoIcon />
-                  <div
-                    className={styles.popoverContent}
-                    style={{ left: '-290px' }}
-                  >
-                    <p>{t('manageProjects:motivationInfo')}</p>
-                    <br />
-                    <p>{t('manageProjects:max300Chars')}</p>
-                  </div>
-                </div>
-              </div>
-            </div>
-          )}
-
-          <div className={styles.formFieldLarge}>
-            <div className={styles.plantingSeasons}>
-              <p className={styles.plantingSeasonsLabel}>
-                {' '}
-                {purpose === 'trees'
-                  ? t('manageProjects:plantingSeasons')
-                  : t('manageProjects:activitySeasons')}{' '}
-              </p>
-              {plantingSeasons.map((month) => {
-                return (
-                  <div
-                    className={styles.multiSelectInput}
-                    key={month.id}
-                    onClick={() => handleSetPlantingSeasons(month.id)}
-                  >
-                    <div
-                      className={`${styles.multiSelectInputCheck} ${
-                        month.isSet ? styles.multiSelectInputCheckTrue : ''
-                      }`}
-                    >
-                      <svg
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="13.02"
-                        height="9.709"
-                        viewBox="0 0 13.02 9.709"
-                      >
-                        <path
-                          id="check-solid"
-                          d="M4.422,74.617.191,70.385a.651.651,0,0,1,0-.921l.921-.921a.651.651,0,0,1,.921,0l2.851,2.85,6.105-6.105a.651.651,0,0,1,.921,0l.921.921a.651.651,0,0,1,0,.921L5.343,74.617a.651.651,0,0,1-.921,0Z"
-                          transform="translate(0 -65.098)"
-                          fill="#fff"
-                        />
-                      </svg>
-                    </div>
-                    <p style={{ color: 'var(--dark)' }}>{month.title}</p>
-                  </div>
-                );
-              })}
-            </div>
-          </div>
-
-          <div className={styles.formField}>
-            <div className={styles.formFieldHalf}>
-              <ThemeProvider theme={materialTheme}>
-                <MuiPickersUtilsProvider
-                  utils={DateFnsUtils}
-                  locale={
-                    localeMapForDate[userLang]
-                      ? localeMapForDate[userLang]
-                      : localeMapForDate['en']
-                  }
-                >
-                  <Controller
-                    render={(properties) => (
-                      <DatePicker
-                        inputRef={register({
-                          required: {
-                            value: true,
-                            message: t(
-                              'manageProjects:employeeCountValidation'
-                            ),
-                          },
-                        })}
-                        label={
-                          purpose === 'trees'
-                            ? t('manageProjects:acquisitionYear')
-                            : t('manageProjects:startingProtectionYear')
-                        }
-                        value={properties.value}
-                        onChange={properties.onChange}
-                        inputVariant="outlined"
-                        TextFieldComponent={MaterialTextField}
-                        autoOk
-                        disableFuture
-                        minDate={new Date(new Date().setFullYear(1950))}
-                        views={['year']}
-                        maxDate={new Date()}
-                      />
-                    )}
-                    name={
-                      purpose === 'trees'
-                        ? 'acquisitionYear'
-                        : 'startingProtectionYear'
-                    }
-                    control={control}
-                    rules={{
-                      required: t('manageProjects:ownershipTypeValidation'),
-                    }}
-                    // defaultValue=""
-                  />
-                  {errors.startingProtectionYear && (
-                    <span className={styles.formErrors}>
-                      {errors.startingProtectionYear.message}
-                    </span>
-                  )}
-                </MuiPickersUtilsProvider>
-              </ThemeProvider>
-            </div>
-            <div style={{ width: '20px' }}></div>
-            {purpose === 'trees' ? (
               <div className={styles.formFieldHalf}>
                 <ThemeProvider theme={materialTheme}>
                   <MuiPickersUtilsProvider
@@ -741,52 +731,93 @@ export default function DetailedAnalysis({
                   </MuiPickersUtilsProvider>
                 </ThemeProvider>
               </div>
-            ) : (
+            </div>
+          ) : (
+            <div
+              className={styles.formFieldLarge}
+              style={{ position: 'relative' }}
+            >
+              <MaterialTextField
+                inputRef={register({
+                  maxLength: {
+                    value: 300,
+                    message: t('manageProjects:max300Chars'),
+                  },
+                })}
+                label={t('manageProjects:ForestProtectionType')}
+                variant="outlined"
+                name="motivation"
+                multiline
+              />
+            </div>
+          )}
+          {purpose === 'trees' ? (
+            <div
+              className={styles.formFieldLarge}
+              style={{ position: 'relative' }}
+            >
+              <MaterialTextField
+                label={t('manageProjects:causeOfDegradation')}
+                variant="outlined"
+                name="degradationCause"
+                multiline
+                inputRef={register({
+                  maxLength: {
+                    value: 300,
+                    message: t('manageProjects:max300Chars'),
+                  },
+                })}
+              />
+              {errors.degradationCause && (
+                <span className={styles.formErrors}>
+                  {errors.degradationCause.message}
+                </span>
+              )}
               <div
-                className={styles.formFieldHalf}
-                style={{ position: 'relative' }}
+                style={{
+                  position: 'absolute',
+                  top: '-9px',
+                  right: '16px',
+                  width: 'fit-content',
+                }}
               >
-                {/* the reason this project has been created (max. 300 characters) */}
-                <MaterialTextField
-                  inputRef={register({
-                    maxLength: {
-                      value: 300,
-                      message: t('manageProjects:max300Chars'),
-                    },
-                  })}
-                  label={t('manageProjects:actions')}
-                  variant="outlined"
-                  name="actions"
-                  multiline
-                />
-
-                <div
-                  style={{
-                    position: 'absolute',
-                    top: '-9px',
-                    right: '16px',
-                    width: 'fit-content',
-                  }}
-                >
-                  <div className={styles.popover}>
-                    <InfoIcon />
-                    <div
-                      className={styles.popoverContent}
-                      style={{ left: '-290px' }}
-                    >
-                      <p>{t('manageProjects:actionInfo')}</p>
-                      <br />
-                      <p>{t('manageProjects:max300Chars')}</p>
-                    </div>
+                <div className={styles.popover}>
+                  <InfoIcon />
+                  <div
+                    className={styles.popoverContent}
+                    style={{ left: '-290px' }}
+                  >
+                    <p>{t('manageProjects:max300Chars')}</p>
                   </div>
                 </div>
               </div>
-            )}
-          </div>
-          <div
-            className={styles.formField}
-            style={{ alignItems: 'flex-start' }}
-          >
+            </div>
+          ) : (
+            <div
+              className={styles.formFieldLarge}
+              style={{ position: 'relative' }}
+            >
+              <MaterialTextField
+                label={t('manageProjects:conservationImpacts')}
+                variant="outlined"
+                name="degradationCause"
+                multiline
+                inputRef={register({
+                  maxLength: {
+                    value: 300,
+                    message: t('manageProjects:max300Chars'),
+                  },
+                })}
+              />
+              {errors.degradationCause && (
+                <span className={styles.formErrors}>
+                  {errors.degradationCause.message}
+                </span>
+              )}
+            </div>
+          )}
+
+          <div className={styles.formField}>
             <div
               className={styles.formFieldHalf}
               style={{ position: 'relative' }}
@@ -830,95 +861,48 @@ export default function DetailedAnalysis({
                 </div>
               </div>
             </div>
-
             <div style={{ width: '20px' }}></div>
-            {purpose === 'trees' ? (
+            <div
+              className={styles.formFieldHalf}
+              style={{ position: 'relative' }}
+            >
+              {/* the reason this project has been created (max. 300 characters) */}
+              <MaterialTextField
+                inputRef={register({
+                  maxLength: {
+                    value: 300,
+                    message: t('manageProjects:max300Chars'),
+                  },
+                })}
+                label={t('manageProjects:whyThisSite')}
+                variant="outlined"
+                name="whyThisSite"
+                multiline
+              />
+              {errors.motivation && (
+                <span className={styles.formErrors}>
+                  {errors.motivation.message}
+                </span>
+              )}
               <div
-                className={styles.formFieldHalf}
-                style={{ position: 'relative' }}
+                style={{
+                  position: 'absolute',
+                  top: '-9px',
+                  right: '16px',
+                  width: 'fit-content',
+                }}
               >
-                {/* the reason this project has been created (max. 300 characters) */}
-                <MaterialTextField
-                  inputRef={register({
-                    maxLength: {
-                      value: 300,
-                      message: t('manageProjects:max300Chars'),
-                    },
-                  })}
-                  label={t('manageProjects:whyThisSite')}
-                  variant="outlined"
-                  name="whyThisSite"
-                  multiline
-                />
-                {errors.motivation && (
-                  <span className={styles.formErrors}>
-                    {errors.motivation.message}
-                  </span>
-                )}
-                <div
-                  style={{
-                    position: 'absolute',
-                    top: '-9px',
-                    right: '16px',
-                    width: 'fit-content',
-                  }}
-                >
-                  <div className={styles.popover}>
-                    <InfoIcon />
-                    <div
-                      className={styles.popoverContent}
-                      style={{ left: '-290px' }}
-                    >
-                      <p>{t('manageProjects:max300Chars')}</p>
-                    </div>
+                <div className={styles.popover}>
+                  <InfoIcon />
+                  <div
+                    className={styles.popoverContent}
+                    style={{ left: '-290px' }}
+                  >
+                    <p>{t('manageProjects:max300Chars')}</p>
                   </div>
                 </div>
               </div>
-            ) : (
-              <div
-                className={styles.formFieldHalf}
-                style={{ position: 'relative' }}
-              >
-                <MaterialTextField
-                  inputRef={register({
-                    required: {
-                      value: true,
-                      message: t('manageProjects:employeeCountValidation'),
-                    },
-                    validate: (value) => parseInt(value, 10) > 0,
-                  })}
-                  label={t('manageProjects:employeeCount')}
-                  variant="outlined"
-                  name="employeeCount"
-                  onInput={(e) => {
-                    e.target.value = e.target.value.replace(/[^0-9]./g, '');
-                  }}
-                />
-                {errors.employeeCount && (
-                  <span className={styles.formErrors}>
-                    {errors.employeeCount.message}
-                  </span>
-                )}
-                <div
-                  style={{
-                    position: 'absolute',
-                    top: '-9px',
-                    right: '16px',
-                    width: 'fit-content',
-                  }}
-                >
-                  <div className={styles.popover}>
-                    <InfoIcon />
-                    <div
-                      className={styles.popoverContent}
-                      style={{ left: '-290px' }}
-                    >
-                      <p>{t('manageProjects:employeesCountInfo')}</p>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            )}
+            </div>
           </div>
 
           <div className={styles.formFieldLarge}>
@@ -962,109 +946,19 @@ export default function DetailedAnalysis({
               })}
             </div>
           </div>
-          {/* <div style={{ width: '20px' }}></div> */}
-          {purpose === 'trees' ? (
-            <div
-              className={styles.formFieldLarge}
-              style={{ position: 'relative' }}
-            >
-              <MaterialTextField
-                label={t('manageProjects:ownerName')}
-                variant="outlined"
-                name="siteOwnerName"
-                inputRef={register()}
-              />
-            </div>
-          ) : (
-            <div
-              className={styles.formFieldLarge}
-              style={{ position: 'relative' }}
-            >
-              <Controller
-                as={
-                  <MaterialTextField
-                    inputRef={register({
-                      required: {
-                        value: true,
-                        message: t('manageProjects:ownershipTypeValidation'),
-                      },
-                    })}
-                    label={t('manageProjects:ownershipType')}
-                    variant="outlined"
-                    select
-                  >
-                    {ownershipType.map((option) => (
-                      <MenuItem
-                        key={option.value}
-                        value={option.value}
-                        classes={{
-                          // option: classes.option,
-                          root: classes.root,
-                        }}
-                      >
-                        {option.label}
-                      </MenuItem>
-                    ))}
-                  </MaterialTextField>
-                }
-                name="ownershipType"
-                rules={{
-                  required: t('manageProjects:ownershipTypeValidation'),
-                }}
-                control={control}
-              />
-              {errors.ownershipType && (
-                <span className={styles.formErrors}>
-                  {errors.ownershipType.message}
-                </span>
-              )}
-            </div>
-          )}
 
-          {purpose === 'trees' ? (
-            <div
-              className={styles.formFieldLarge}
-              style={{ position: 'relative' }}
-            >
-              <MaterialTextField
-                label={t('manageProjects:causeOfDegradation')}
-                variant="outlined"
-                name="degradationCause"
-                multiline
-                inputRef={register({
-                  maxLength: {
-                    value: 300,
-                    message: t('manageProjects:max300Chars'),
-                  },
-                })}
-              />
-              {errors.degradationCause && (
-                <span className={styles.formErrors}>
-                  {errors.degradationCause.message}
-                </span>
-              )}
-              <div
-                style={{
-                  position: 'absolute',
-                  top: '-9px',
-                  right: '16px',
-                  width: 'fit-content',
-                }}
-              >
-                <div className={styles.popover}>
-                  <InfoIcon />
-                  <div
-                    className={styles.popoverContent}
-                    style={{ left: '-290px' }}
-                  >
-                    <p>{t('manageProjects:max300Chars')}</p>
-                  </div>
-                </div>
-              </div>
-            </div>
-          ) : (
-            <></>
-          )}
+          <div
+            className={styles.formFieldLarge}
+            style={{ position: 'relative' }}
+          >
+            <MaterialTextField
+              label={t('manageProjects:ownerName')}
+              variant="outlined"
+              name="siteOwnerName"
+              inputRef={register()}
+            />
+          </div>
+
           <div
             className={styles.formFieldLarge}
             style={{ position: 'relative' }}
@@ -1086,27 +980,50 @@ export default function DetailedAnalysis({
                 {errors.longTermPlan.message}
               </span>
             )}
-            <div
-              style={{
-                position: 'absolute',
-                top: '-9px',
-                right: '16px',
-                width: 'fit-content',
-              }}
-            >
-              <div className={styles.popover}>
-                <InfoIcon />
-                <div
-                  className={styles.popoverContent}
-                  style={{ left: '-290px' }}
-                >
-                  <p>{t('manageProjects:longTermPlanInfo')}</p>
-                  <br />
-                  <p>{t('manageProjects:max300Chars')}</p>
-                </div>
-              </div>
-            </div>
           </div>
+          {purpose === 'conservation' ? (
+            <div className={styles.formFieldHalf}>
+              <Controller
+                as={
+                  <MaterialTextField
+                    inputRef={register({
+                      required: {
+                        value: true,
+                        message: t(
+                          'manageProjects:endangeredSpeciesValidation'
+                        ),
+                      },
+                    })}
+                    label={t('manageProjects:endangeredSpecies')}
+                    variant="outlined"
+                    select
+                  >
+                    {enSpecies.map((option) => (
+                      <MenuItem
+                        key={option.id}
+                        value={option.id}
+                        classes={{ root: classes.root }}
+                      >
+                        {option.label}
+                      </MenuItem>
+                    ))}
+                  </MaterialTextField>
+                }
+                name={t('manageProjects:endangeredSpecies')}
+                rules={{
+                  required: t('manageProjects:endangeredSpeciesValidation'),
+                }}
+                control={control}
+              />
+              {errors.degradationCause && (
+                <span className={styles.formErrors}>
+                  {errors.degradationCause.message}
+                </span>
+              )}
+            </div>
+          ) : (
+            <></>
+          )}
 
           <ProjectCertificates
             projectGUID={projectGUID}
@@ -1114,48 +1031,51 @@ export default function DetailedAnalysis({
             setIsUploadingData={setIsUploadingData}
             userLang={userLang}
           />
-        </div>
+          {errorMessage && errorMessage !== '' ? (
+            <div className={styles.formFieldLarge}>
+              <h4 className={styles.errorMessage}>{errorMessage}</h4>
+            </div>
+          ) : null}
 
-        {errorMessage && errorMessage !== '' ? (
-          <div className={styles.formFieldLarge}>
-            <h4 className={styles.errorMessage}>{errorMessage}</h4>
-          </div>
-        ) : null}
-
-        <div className={styles.formField} style={{ marginTop: '48px' }}>
-          <div className={`${styles.formFieldHalf}`}>
-            <button
-              onClick={handleBack}
-              className="secondaryButton"
-              style={{ width: '234px', height: '46px', marginLeft: '45px' }}
-            >
-              <BackArrow />
-              <p>{t('manageProjects:backToMedia')}</p>
-            </button>
-          </div>
-          <div style={{ width: '20px' }}></div>
-          <div className={`${styles.formFieldHalf}`}>
-            <button
-              onClick={handleSubmit(onSubmit)}
-              className="primaryButton"
-              style={{ width: '169px', height: '46px', marginRight: '20px' }}
-              data-test-id="detailAnalysisCont"
-            >
-              {isUploadingData ? (
-                <div className={styles.spinner}></div>
-              ) : (
-                t('manageProjects:saveAndContinue')
-              )}
-            </button>
-          </div>
-          <div className={`${styles.formFieldHalf}`}>
-            <button
-              className="primaryButton"
-              style={{ width: '89px', marginRight: '45px' }}
-              onClick={handleNext}
-            >
-              {t('manageProjects:skip')}
-            </button>
+          <div className={styles.formFieldLarge} style={{ marginTop: '48px' }}>
+            <div className={`${styles.formFieldHalf}`}>
+              <button
+                onClick={handleBack}
+                className="secondaryButton"
+                style={{ width: '234px', height: '46px', marginLeft: '45px' }}
+              >
+                <BackArrow />
+                <p>{t('manageProjects:backToMedia')}</p>
+              </button>
+            </div>
+            <div style={{ width: '20px' }}></div>
+            <div className={styles.formFieldHalf}>
+              <button
+                onClick={handleSubmit(onSubmit)}
+                className="primaryButton"
+                style={{
+                  width: '169px',
+                  height: '46px',
+                  marginRight: '20px',
+                }}
+                data-test-id="detailAnalysisCont"
+              >
+                {isUploadingData ? (
+                  <div className={styles.spinner}></div>
+                ) : (
+                  t('manageProjects:saveAndContinue')
+                )}
+              </button>
+            </div>
+            <div className={`${styles.formFieldHalf}`}>
+              <button
+                className="primaryButton"
+                style={{ width: '89px', marginRight: '45px' }}
+                onClick={handleNext}
+              >
+                {t('manageProjects:skip')}
+              </button>
+            </div>
           </div>
         </div>
       </form>
